requirements.txt
Flask==2.3.3
python-whois==0.8.0
requests==2.31.0